Wrap restGet Calls in Arrays
Without doing this, the server itself must wrap the JSON in brackets—which is not documented or intuitive—and, as an example of why this is required, Vortex to throw an error attempting to parse the request without this, skipping the server and querying the next one.

Do note that using data from meta servers throws other, more noticeable errors in Vortex at the moment. This is just the first domino.