support for "in-process" meta data server
using a callback function instead of a network port to communicate with

this always fixes an issue if a mod is initially stored with the md5 hash
of the file unknown and later, once the file is actually downloaded, we
do have the actual hash. Previously that caused issues because we had
the lookup by name stored with the wrong md5 hash.